rename sessions analytics package
