+++ internal/zero/telemetry/sessions/activeusers.go
@@ -1,4 +1,4 @@
-package analytics
+package sessions
 
 import (
 	"time"
