+++ internal/zero/telemetry/sessions/sessions.go
@@ -1,4 +1,4 @@
-package analytics
+package sessions
 
 import (
 	"context"
